# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  library-management-system-go_internal_dto.CreateBookRequest:
    properties:
      author:
        type: string
      category:
        type: string
      isbn:
        type: string
      publication_year:
        type: integer
      title:
        type: string
      total_copies:
        minimum: 1
        type: integer
    required:
    - author
    - category
    - isbn
    - publication_year
    - title
    - total_copies
    type: object
...
  library-management-system-go_internal_dto.CreateFineRequest:
    properties:
      amount:
        type: number
      due_date:
        type: string
      lending_id:
        type: integer
      notes:
        type: string
      reason:
        type: string
    required:
    - amount
    - due_date
    - lending_id
    - reason
    type: object
  library-management-system-go_internal_dto.CreateLendingRequest:
    properties:
      book_id:
        type: integer
      due_date:
        type: string
      max_renewals:
        type: integer
      notes:
        type: string
    required:
    - book_id
    - due_date
    type: object
  library-management-system-go_internal_dto.CreateReservationRequest:
    properties:
      book_id:
        type: integer
      expiry_date:
        type: string
      notes:
        type: string
    required:
    - book_id
    - expiry_date
    type: object
...
  library-management-system-go_internal_dto.UpdateBookRequest:
    properties:
      author:
        type: string
      category:
        type: string
      isbn:
        type: string
      publication_year:
        type: integer
      status:
        type: string
      title:
        type: string
      total_copies:
        type: integer
    type: object
...
  library-management-system-go_internal_dto.UpdateLendingRequest:
    properties:
      due_date:
        type: string
      max_renewals:
        type: integer
      notes:
        type: string
      renewal_count:
        type: integer
      return_date:
        type: string
      status:
        type: string
    type: object
  library-management-system-go_internal_dto.UpdateReservationRequest:
    properties:
      cancelled_date:
        type: string
      fulfilled_date:
        type: string
      lending_id:
        type: integer
      notes:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for school library management.
  title: School Library Management API
  version: "1.0"
paths:
...
  /books:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List books
      tags:
      - books
    post:
      consumes:
      - application/json
      parameters:
      - description: Create book request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create book
      tags:
      - books
  /books/{id}:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete book
      tags:
      - books
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get book
      tags:
      - books
    put:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update book request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update book
      tags:
      - books
...
  /fines:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List fines
      tags:
      - fines
    post:
      consumes:
      - application/json
      parameters:
      - description: Create fine request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateFineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create fine
      tags:
      - fines
  /fines/{id}:
    get:
      parameters:
      - description: Fine ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get fine
      tags:
      - fines
  /lendings:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List lendings
      tags:
      - lendings
    post:
      consumes:
      - application/json
      parameters:
      - description: Create lending request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateLendingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create lending
      tags:
      - lendings
  /lendings/{id}:
    delete:
      parameters:
      - description: Lending ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete lending
      tags:
      - lendings
    get:
      parameters:
      - description: Lending ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get lending
      tags:
      - lendings
    put:
      consumes:
      - application/json
      parameters:
      - description: Lending ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update lending request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateLendingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update lending
      tags:
      - lendings
...
  /reservations:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      parameters:
      - description: Create reservation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create reservation
      tags:
      - reservations
  /reservations/{id}:
    delete:
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete reservation
      tags:
      - reservations
    get:
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get reservation
      tags:
      - reservations
    put:
      consumes:
      - application/json
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update reservation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/library-management-system-go_internal_dto.UpdateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update reservation
      tags:
      - reservations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"